Refactoring
The original class was just too big to handle in one file. Breaking it
up a bit to make it easier to manage.
