+++ src/main/java/com/github/seqware/OxoGWrapperWorkflow.java
@@ -7,264 +7,14 @@ import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.nio.file.StandardOpenOption;
-import java.util.ArrayList;
-import java.util.List;
 import java.util.stream.Collectors;
 
-import net.sourceforge.seqware.pipeline.workflowV2.AbstractWorkflowDataModel;
 import net.sourceforge.seqware.pipeline.workflowV2.model.Job;
 
-public class OxoGWrapperWorkflow extends AbstractWorkflowDataModel {
+public class OxoGWrapperWorkflow extends BaseOxoGWrapperWorkflow {
 
-	//ugh... so many fields. There's probably a better way to do this, just no time right now.
-	private String oxoQScore = "";
-	// private String donorID;
-	private String aliquotID;
-	private String bamNormalObjectID;
-	private String bamNormalIndexObjectID;
-	private String bamTumourObjectID;
-	private String bamTumourIndexObjectID;
+
 	
-	private String sangerSNVVCFObjectID;
-	private String dkfzemblSNVVCFObjectID;
-	private String broadSNVVCFObjectID;
-	private String museSNVVCFObjectID;
-
-	private String sangerIndelVCFObjectID;
-	private String dkfzemblIndelVCFObjectID;
-	private String broadIndelVCFObjectID;
-
-	private String sangerSVVCFObjectID;
-	private String dkfzemblSVVCFObjectID;
-	private String broadSVVCFObjectID;
-
-	private String sangerSNVIndexObjectID;
-	private String dkfzemblSNVIndexObjectID;
-	private String broadSNVIndexObjectID;
-	private String museSNVIndexObjectID;
-
-	private String sangerIndelIndexObjectID;
-	private String dkfzemblIndelIndexObjectID;
-	private String broadIndelIndexObjectID;
-
-	private String sangerSVIndexObjectID;
-	private String dkfzemblSVIndexObjectID;
-	private String broadSVIndexObjectID;
-	
-	private String uploadURL;
-
-	private String JSONrepo = null;
-	private String JSONrepoName = "oxog-ops";
-	private String JSONfolderName = null;
-	private String JSONlocation = "/datastore/gitroot";
-	private String JSONfileName = null;
-
-	private String GITemail = "";
-	private String GITname = "ICGC AUTOMATION";
-	private String GITPemFile = "";
-
-	//These will be needed so that vcf-uploader can generate the analysis.xml and manifest.xml files
-	private String tumourMetdataURL;
-	private String normalMetdataURL;
-	
-	private String tumourBAMFileName;
-	private String normalBAMFileName;
-	private String sangerSNVName;
-	private String dkfzEmblSNVName;
-	private String broadSNVName;
-	private String museSNVName;
-
-	private String sangerSVName;
-	private String dkfzEmblSVName;
-	private String broadSVName;
-
-	private String sangerIndelName;
-	private String dkfzEmblIndelName;
-	private String broadIndelName;
-	
-	private int snvPadding = 10;
-	private int svPadding = 10;
-	private int indelPadding = 200;
-	
-	private String storageSource = "collab";
-	
-	private boolean gitMoveTestMode = false;
-	
-	private List<String> filesToUpload = new ArrayList<String>();
-	
-	//Paths to VCFs generated by merging types across workflows. 
-	private String snvVCF;
-	private String svVCF;
-	private String indelVCF;
-	private String sangerNormalizedIndelVCFName;
-	private String broadNormalizedIndelVCFName;
-	private String dkfzEmblNormalizedIndelVCFName;
-	//private String museNormalizedIndelVCFName;
-
-	private String sangerExtractedSNVVCFName;
-	private String broadExtractedSNVVCFName;
-	private String dkfzEmblExtractedSNVVCFName;
-	//private String museExtractedSNVVCFName;
-
-	private String sangerGnosID;
-	private String broadGnosID;
-	private String dkfzemblGnosID;
-	private String museGnosID;
-	
-	//This could be used implement a sort of local-file mode. For now, it's just used to speed up testing.
-	private boolean skipDownload = false;
-	private boolean skipUpload = false;
-	
-	//Path to reference file usd for normalization, *relative* to /datastore/refdata
-	private String refFile = "pcawg/genome.fa";
-	private String tumourBamGnosID;
-	private String normalBamGnosID;
-	private String uploadKey;
-	
-	//These two variables can be used to skip running OxoG and variant bam, in case they have already been run.
-	//Intended to speed up testing by skipping these steps when results already exist.
-	private boolean skipOxoG = false;
-	private boolean skipVariantBam = false;
-	
-	/**
-	 * Get a property name that is mandatory
-	 * @param propName The name of the property
-	 * @return The property, as a String. Convert to other types if you need to.
-	 * @throws Exception An Exception with the message "Property with key <i>propName</i> cannot be null" will be thrown if property is not found.
-	 */
-	private String getMandatoryProperty(String propName) throws Exception
-	{
-		if (hasPropertyAndNotNull(propName)) {
-			return getProperty(propName);
-		}
-		else {
-			throw new Exception ("Property with key "+propName+ " cannot be null!");
-		}
-	}
-
-	/**
-	 * Initial setup.
-	 */
-	private void init() {
-		try {
-			
-			this.oxoQScore = this.getMandatoryProperty(JSONUtils.OXOQ_SCORE);
-			this.JSONrepo = this.getMandatoryProperty("JSONrepo");
-			this.JSONrepoName = this.getMandatoryProperty("JSONrepoName");
-			this.JSONfolderName = this.getMandatoryProperty("JSONfolderName");
-			this.JSONfileName = this.getMandatoryProperty("JSONfileName");
-			this.GITemail = this.getMandatoryProperty("GITemail");
-			this.GITname = this.getMandatoryProperty("GITname");
-			
-			this.bamNormalObjectID = this.getMandatoryProperty(JSONUtils.BAM_NORMAL_OBJECT_ID);
-			this.normalMetdataURL = this.getMandatoryProperty(JSONUtils.BAM_NORMAL_METADATA_URL);
-			this.bamTumourObjectID = this.getMandatoryProperty(JSONUtils.BAM_TUMOUR_OBJECT_ID);
-			this.tumourMetdataURL = this.getMandatoryProperty(JSONUtils.BAM_TUMOUR_METADATA_URL);
-			this.sangerSNVVCFObjectID = this.getMandatoryProperty(JSONUtils.SANGER_SNV_VCF_OBJECT_ID);
-			this.dkfzemblSNVVCFObjectID = this.getMandatoryProperty(JSONUtils.DKFZEMBL_SNV_VCF_OBJECT_ID);
-			this.broadSNVVCFObjectID = this.getMandatoryProperty(JSONUtils.BROAD_SNV_VCF_OBJECT_ID);
-			this.museSNVVCFObjectID = this.getMandatoryProperty(JSONUtils.MUSE_VCF_OBJECT_ID);
-			this.uploadURL = this.getMandatoryProperty("uploadURL");
-			this.aliquotID = this.getMandatoryProperty(JSONUtils.ALIQUOT_ID);
-			
-			this.GITPemFile = this.getMandatoryProperty("GITPemFile");
-
-			this.normalBAMFileName = this.getMandatoryProperty(JSONUtils.BAM_NORMAL_FILE_NAME);
-			this.tumourBAMFileName = this.getMandatoryProperty(JSONUtils.BAM_TUMOUR_FILE_NAME);
-			
-			this.sangerSNVName = this.getMandatoryProperty(JSONUtils.SANGER_SNV_VCF_NAME);
-			this.broadSNVName = this.getMandatoryProperty(JSONUtils.BROAD_SNV_VCF_NAME);
-			this.dkfzEmblSNVName = this.getMandatoryProperty(JSONUtils.DKFZEMBL_SNV_VCF_NAME);
-			this.museSNVName = this.getMandatoryProperty(JSONUtils.MUSE_VCF_NAME);
-			
-			this.sangerIndelName = this.getMandatoryProperty(JSONUtils.SANGER_INDEL_VCF_NAME);
-			this.dkfzEmblIndelName = this.getMandatoryProperty(JSONUtils.DKFZEMBL_INDEL_VCF_NAME);
-			this.broadIndelName = this.getMandatoryProperty(JSONUtils.BROAD_INDEL_VCF_NAME);
-
-			this.sangerSVName = this.getMandatoryProperty(JSONUtils.SANGER_SV_VCF_NAME);
-			this.dkfzEmblSVName = this.getMandatoryProperty(JSONUtils.DKFZEMBL_SV_VCF_NAME);
-			this.broadSVName = this.getMandatoryProperty(JSONUtils.BROAD_SV_VCF_NAME);
-
-			this.sangerSVVCFObjectID = this.getMandatoryProperty(JSONUtils.SANGER_SV_VCF_OBJECT_ID);
-			this.broadSVVCFObjectID = this.getMandatoryProperty(JSONUtils.BROAD_SV_VCF_OBJECT_ID);
-			this.dkfzemblSVVCFObjectID = this.getMandatoryProperty(JSONUtils.DKFZEMBL_SV_VCF_OBJECT_ID);
-			
-			this.sangerIndelVCFObjectID = this.getMandatoryProperty(JSONUtils.SANGER_INDEL_VCF_OBJECT_ID);
-			this.broadIndelVCFObjectID = this.getMandatoryProperty(JSONUtils.BROAD_INDEL_VCF_OBJECT_ID);
-			this.dkfzemblIndelVCFObjectID = this.getMandatoryProperty(JSONUtils.DKFZEMBL_INDEL_VCF_OBJECT_ID);
-			
-			this.sangerSNVIndexObjectID = this.getMandatoryProperty(JSONUtils.SANGER_SNV_INDEX_OBJECT_ID);
-			this.broadSNVIndexObjectID = this.getMandatoryProperty(JSONUtils.BROAD_SNV_INDEX_OBJECT_ID);
-			this.dkfzemblSNVIndexObjectID = this.getMandatoryProperty(JSONUtils.DKFZEMBL_SNV_INDEX_OBJECT_ID);
-			this.museSNVIndexObjectID = this.getMandatoryProperty(JSONUtils.MUSE_SNV_INDEX_OBJECT_ID);
-			
-			this.sangerSVIndexObjectID = this.getMandatoryProperty(JSONUtils.SANGER_SNV_INDEX_OBJECT_ID);
-			this.broadSVIndexObjectID = this.getMandatoryProperty(JSONUtils.BROAD_SV_INDEX_OBJECT_ID);
-			this.dkfzemblSVIndexObjectID = this.getMandatoryProperty(JSONUtils.DKFZEMBL_SV_INDEX_OBJECT_ID);
-			
-			this.sangerIndelIndexObjectID = this.getMandatoryProperty(JSONUtils.SANGER_INDEL_INDEX_OBJECT_ID);
-			this.broadIndelIndexObjectID = this.getMandatoryProperty(JSONUtils.BROAD_INDEL_INDEX_OBJECT_ID);
-			this.dkfzemblIndelIndexObjectID = this.getMandatoryProperty(JSONUtils.DKFZEMBL_INDEL_INDEX_OBJECT_ID);
-			
-			this.bamNormalIndexObjectID = this.getMandatoryProperty(JSONUtils.BAM_NORMAL_INDEX_OBJECT_ID);
-			this.bamTumourIndexObjectID = this.getMandatoryProperty(JSONUtils.BAM_TUMOUR_INDEX_OBJECT_ID);
-
-			this.sangerGnosID = this.getMandatoryProperty(JSONUtils.SANGER_GNOS_ID);
-			this.broadGnosID = this.getMandatoryProperty(JSONUtils.BROAD_GNOS_ID);
-			this.dkfzemblGnosID = this.getMandatoryProperty(JSONUtils.DKFZEMBL_GNOS_ID);
-			this.museGnosID = this.getMandatoryProperty(JSONUtils.MUSE_GNOS_ID);
-			
-			this.normalBamGnosID= this.getMandatoryProperty(JSONUtils.BAM_NORMAL_GNOS_ID);
-			this.tumourBamGnosID= this.getMandatoryProperty(JSONUtils.BAM_TUMOUR_GNOS_ID);
-			
-			this.uploadKey= this.getMandatoryProperty("uploadKey");
-			
-			if (hasPropertyAndNotNull("gitMoveTestMode")) {
-				this.gitMoveTestMode = Boolean.valueOf(getProperty("gitMoveTestMode"));
-			}
-			
-			if (hasPropertyAndNotNull("storageSource")) {
-				this.storageSource = getProperty("storageSource");
-			}
-			
-			if (hasPropertyAndNotNull("snvPadding")) {
-				this.snvPadding = Integer.valueOf(getProperty("snvPadding"));
-			}
-			
-			if (hasPropertyAndNotNull("svPadding")) {
-				this.svPadding = Integer.valueOf(getProperty("svPadding"));
-			}
-			
-			if (hasPropertyAndNotNull("indelPadding")) {
-				this.indelPadding = Integer.valueOf(getProperty("indelPadding"));
-			}
-			
-			if (hasPropertyAndNotNull("refFile")) {
-				this.refFile = getProperty("refFile");
-			}
-			
-			if (hasPropertyAndNotNull("skipDownload")) {
-				this.skipDownload = Boolean.valueOf(getProperty("skipDownload"));
-			}
-			
-			if (hasPropertyAndNotNull("skipUpload")) {
-				this.skipUpload = Boolean.valueOf(getProperty("skipUpload"));
-			}
-			
-			if (hasPropertyAndNotNull("skipOxoG")) {
-				this.skipOxoG = Boolean.valueOf(getProperty("skipOxoG"));
-			}
-			
-			if (hasPropertyAndNotNull("skipVariantBam")) {
-				this.skipVariantBam = Boolean.valueOf(getProperty("skipVariantBam"));
-			}
-			
-			this.generateRulesFile();
-			
-		} catch (Exception e) {
-			throw new RuntimeException("Exception encountered during workflow init: "+e.getMessage(),e);
-		}
-	}
 
 	/**
 	 * Generates a rules file that is used for the variant program that produces minibams.
@@ -838,6 +588,7 @@ public class OxoGWrapperWorkflow extends AbstractWorkflowDataModel {
 	public void buildWorkflow() {
 		try {
 			this.init();
+			this.generateRulesFile();
 			// Pull the repo.
 			Job configJob = GitUtils.gitConfig(this.getWorkflow(), this.GITname, this.GITemail);
